backend/requirements.txt
fastapi
uvicorn
mediapipe==0.10.30
opencv-python
scikit-learn
numpy
python-multipart
Pillow
requests
openai
python-dotenv
